Enable additional plugins for Kyma organizations
This is small feat. which enables some of the plugins we tested on the test-infra for all orgs.This will not affect the usual workflow. Generally these changes allow to add labels based on OWNERS files, use lifecycle plugin which recently had updates to work with stalebot and verify owners files.