feat(external_secrets.yaml): add ExternalSecret resources for signify-dev-secret and sa-kyma-push-images This commit adds two ExternalSecret resources to the Kubernetes cluster. The signify-dev-secret ExternalSecret is used to sync secrets for the signify-dev-secret GCP secret. The sa-kyma-push-images ExternalSecret is used to sync secrets for the sa-dev-kyma-project GCP secret. Both ExternalSecret resources use the gcp-secretstore ClusterSecretStore and have a refresh interval of 10 minutes. The deletion policy is set to "Delete" for both ExternalSecret resources.
chore(external-secrets): update values.yaml
This commit updates the values.yaml file for the external-secrets deployment. It changes the name of the service account to 'tekton-external-secrets' for consistency with the naming conventions. The 'prometheus' section is commented out as it is deprecated and will be removed in the future. The 'webhook' and 'certController' sections now have a name specified for the service account, which improves readability and consistency.