feat(golangci-lint): add golanci-lint Task. The golangci-lint Task is a Tekton task that runs a golang linter. It is designed to be triggered by prow as a prowjob. The task uses the golangci-lint tool.
Improve task description and workspace description.

The task description has been updated to include information about the prowjob trigger and the use of the golangci-lint tool.
The workspace description has been updated to include information about the default mount path and the fact that it should be a shared workspace.
Information about the golangci-lint tool and its configuration has been added to the task description.

Add sample prowjob for linting go code.