Tekton go unit test task (#7302)
* Task first version.

* Added unit-test-go tekton task.

* feat(unit-test-go): add unit test task for Go code
This commit adds a new task for running unit tests for Go code. The task is tailored to be triggered by prow as a prowjob and uses the golangci-lint tool to run the tests. The task is compatible with Tekton v0.36.0 and above and can be installed using the provided kubectl command. The task requires a workspace named "repo" where the sources to test are stored. The task has optional parameters for specifying the path to the Go code to test, as well as the GitHub org and repo that triggered the prowjob. The commit also includes a sample presubmit prowjob configuration for running the unit tests and a CONTRIBUTING document.

feat(prowjob-testing-go-code.yaml): add tekton pipeline for running go unit tests
This commit adds a new file `prowjob-testing-go-code.yaml` which contains a Tekton pipeline for running Go unit tests. The pipeline is triggered by a presubmit event and uses the `tekton-pipeline` agent. The pipeline clones the repository refs provided by the prowjob and runs the Go unit tests. The pipeline is run on a dedicated Tekton cluster and uses a premium-rwo storage class with a 1Gi storage request.

refactor(unit-test-go.yaml): rename task name from unit-tests-go to go-unit-tests
The task name has been changed from unit-tests-go to go-unit-tests to improve consistency with naming conventions.

* Removed results. They were not set by a task.
Removed env vars. They were not used by step.

* Env vars are required by some tests.

* Aligned names.

* Apply suggestions from code review

Co-authored-by: Iwona Langer <iwona.langer@sap.com>

* Removed wrong information about used tool from description.
This is a Task not Pipeline. Corrected this in Platform section.

* Fixed naming inconsistency in task description and readme.

* Apply suggestions from code review

Co-authored-by: Iwona Langer <iwona.langer@sap.com>

* Addressed review comment.

---------

Co-authored-by: Iwona Langer <iwona.langer@sap.com>