Clonerefs and image-builder tasks for tekton (#7258)
* feat(prow): add tekton-demo presubmit job
This commit adds a new presubmit job called tekton-demo to the kyma-project/kyma repository. The job runs on the tekton-pipeline agent and cluster, and uses a Tekton pipeline run spec to define the workspaces and pipeline tasks. The pipeline tasks include clonerefs and list-files, which clone the repository and list the files in the /home/prow/go directory, respectively. The job is not decorated and is set to always_run: false.

feat(build-image.yaml): add build-image task to Tekton catalog
The build-image task is added to the Tekton catalog. This task clones a repo from the provided URL into the output workspace and supports sparse checkouts. It also has parameters for name, config, context, and dockerfile. The task uses the image-builder container from the eu.gcr.io/sap-kyma-neighbors-dev repository. The task requires 1.5Gi of memory and 1 CPU.

feat(clone-refs.yaml): add clone-refs task to Tekton catalog
The clone-refs task is a Git task that clones a repository from a provided URL into the output workspace. It supports cloning into a subdirectory and sparse checkouts. This task is added to the Tekton catalog with workspaces for logs, repo, and tmp. The task is executed by running the clone-refs step, which uses the europe-docker.pkg.dev/kyma-project/dev/test-infra/tekton-clonerefs:latest image and sets environment variables for JOB_SPEC, LOG, and SRC_ROOT.

feat(clone-refs): add Dockerfile for clonerefs Tekton task
This commit adds a Dockerfile for the clonerefs Tekton task. The Dockerfile uses a multi-stage build to first build the clonerefs binary using golang:1.20.2-alpine3.17 as the builder image. The final image is based on alpine:3.17.2 and includes the clonerefs binary, ca-certificates, and git. The entrypoint for the image is set to /clonerefs.

feat(clone-refs): add support for extra_refs in job spec
This commit adds support for extra_refs in the job spec. If present, the extra_refs are unmarshalled and appended to the gitRefs slice. The clonerefs.Options struct is updated to include the new GitRefs field. The Validate and Run methods are updated to handle the new field. If the extra_refs field is not present, the refs field is used instead.

feat(external-secrets): add external secrets and secret store YAML files
This commit adds two new YAML files for external secrets and secret store. The external_secrets.yaml file defines an ExternalSecret resource that references a ClusterSecretStore resource named gcp-secretstore. The secrets_store.yaml file defines the ClusterSecretStore resource named gcp-secretstore, which uses the gcpsm provider and specifies the projectID as sap-kyma-prow.

feat(external-secrets): add values.yaml for external-secrets Helm chart
This commit adds the values.yaml file for the external-secrets Helm chart. The file contains various configuration options for the chart, including the number of replicas, image repository and pull policy, CRD installation, leader election, controller class, scoped namespace, and more. The file also includes options for creating a service account, RBAC resources, and a pod disruption budget. Additionally, the file includes options for creating a webhook deployment and a certificate controller deployment, as well as options for configuring Prometheus metrics and DNS options.

* feat(external_secrets.yaml): add ExternalSecret resources for signify-dev-secret and sa-kyma-push-images
This commit adds two ExternalSecret resources to the Kubernetes cluster. The signify-dev-secret ExternalSecret is used to sync secrets for the signify-dev-secret GCP secret. The sa-kyma-push-images ExternalSecret is used to sync secrets for the sa-dev-kyma-project GCP secret. Both ExternalSecret resources use the gcp-secretstore ClusterSecretStore and have a refresh interval of 10 minutes. The deletion policy is set to "Delete" for both ExternalSecret resources.

chore(external-secrets): update values.yaml
This commit updates the values.yaml file for the external-secrets deployment. It changes the name of the service account to 'tekton-external-secrets' for consistency with the naming conventions. The 'prometheus' section is commented out as it is deprecated and will be removed in the future. The 'webhook' and 'certController' sections now have a name specified for the service account, which improves readability and consistency.

* "Add new Task for OCI image build"

* Add wrapper for k8s test-infra clonerefs to run in as Tekton pipeline
task.
Dockerfile for building OCI image to use in Task.

* clone-refs task definition.
Moved task def to location supported by tekton hub.

* Example prowjob using tekton pipeline.

* Added prowjob pub sub config.

* Added readme and sample prowjob.

* Added readme files and sample prowjobs using tekton pipeline.
Improved descriptions and naming.

* Fixed typo.

* Aligned workspace and task names.

* Address review comments.
Fixed mistakes in docs and paths.
Moved tekton-demo.yaml to docs/examples.

* This is a task documentation. It's not requried to say it in the title. Removed a Task word to align with other task documentation.

* Aligned file name with documentation.

* Apply suggestions from code review

Co-authored-by: Iwona Langer <iwona.langer@sap.com>

* Addressed review comments.

---------

Co-authored-by: Iwona Langer <iwona.langer@sap.com>