Gatekeeper workload clusters manifests (#7521)
* Add presubmit job for gatekeeper-poc

feat(Gatekeeper): add ConstraintTemplate for controling usage of restricted secrets
This commit adds a new constraint template for Gatekeeper that controls any Pod's ability to use restricted secrets. The constraint template defines a set of criteria that must be met for a Pod to be allowed to use a restricted secret. The criteria include labels, restricted secrets, trusted service accounts, trusted images, trusted commands and trusted arguments. The commit also includes a Rego policy that checks if a container is using a restricted secret and does not match trusted usage criteria. The Rego policy also checks if the pod is matching trusted usage criteria, which are defined in the ConstraintTemplate parameters. The Rego policy uses several functions to check if the pod has required labels, is using a trusted image, is using trusted commands, and is using trusted arguments. If the criteria are not met, a violation message is generated.

feat(Gatekeeper): add ConstraintTemplate for controlling usage of restricted service accounts
This commit adds a new ConstraintTemplate for controlling a Kubernetes workload's ability to use restricted service accounts. The ConstraintTemplate defines a CRD named ServiceAccountTrustedUsage, which has validation rules for labels, restricted service accounts, trusted images, trusted commands and trusted arguments. The ConstraintTemplate also includes a Rego policy that checks if a pod is using a restricted service account and does not match trusted usage criteria. The Rego policy also checks if the pod is matching trusted usage criteria, which are defined in the ConstraintTemplate parameters. The Rego policy uses several functions to check if the pod has required labels, is using a trusted image, is using trusted commands, and is using trusted arguments. If the criteria are not met, a violation message is generated.

feat(Gatekeeper): add test data for trusted service account usage constraint template
This commit adds test data for the trusted service account usage constraint template. The test data includes a sample pod object with two containers, one of which uses a trusted image and the other uses an untrusted image. This test data can be used to test the trustedSAUsage constraint template in opa playground.

feat(Gatekeeper): add test data for trustedSecretUsage constraint template
This commit adds test data for the trustedSecretUsage constraint template. The test data includes parameters for labels, restricted secrets, trusted service accounts, and trusted images. The review object includes metadata, spec, containers, initContainers, and volumes. This test data can be used to test the trustedSecretUsage constraint template in opa playground.

chore(tekton): update default service account name
The default service account name has been changed from "tekton" to "default" in the config_defaults.yaml file. This change fix security issue as "tekton" service account which was used to run Tasks is bind to cluster-admin role. This role is not intended to be used as Task identity and has to many permissions.

feat(gatekeeper-constraints): add tekton-signify-dev-secret-trusted-usage constraint
This commit adds a new constraint to the gatekeeper-constraints directory. The constraint is named tekton-signify-dev-secret-trusted-usage and is of type SecretTrustedUsage. It specifies a match for Pods and restricts the usage of the signify-dev-secret secret to only Pods with the prow.k8s.io/type label set to "presubmit". It also specifies a trusted image with the image-builder entrypoint and arguments.

feat(gatekeeper-constraints): add tekton-signify-prod-secret-trusted-usage constraint
This commit adds a new constraint to the gatekeeper-constraints directory. The constraint is named tekton-signify-prod-secret-trusted-usage and is of type SecretTrustedUsage. It specifies a match for Pods and includes parameters for labels, restrictedSecrets, and trustedImages. The constraint is designed to restrict the use of the signify-prod-secret secret to only Pods with the prow.k8s.io/type label set to "postsubmit" and with trusted images that match the specified image and command/args.

feat(gatekeeper-constraints): add tekton-image-builder-sa-trusted-usage constraint
A new constraint has been added to the gatekeeper-constraints directory which specifies trusted images for the image-builder service account. This constraint ensures that only trusted images can run as image-builder service account. It specifies a match for Pods and restricts the usage of the image-builder service account to only Pods that use a trusted image with the trusted entrypoint and arguments.

feat(tekton): add RBAC rules for secure access to signify dev and prod secrets
This commit adds RBAC rules to allow only trusted service accounts to access signify dev and prod secrets over the k8s API. The rules allow the service accounts image-builder, tekton-pipelines-controller, and external-secrets to update, get, list, watch, patch, create, and delete the secrets. A new service account image-builder is also created in the default namespace with automountServiceAccountToken set to true.

* Added basic terraform dir structure.
Added default stable gatekeeper manifest file.

* Draft module and dev env.

* Added terraform config for terraform-executor and gatekeeper modules and environments.

* Using terraform resources for authentication in k8s clusters.
Consolidated prod environments under one kyma-test-infra-prod environment.

* Typos in file names.

* Removed unused files.

* Use tfcmt config file to patch comments. This way it's not needed to add cli param for plan actions.

* Test with disabled loading config.

* Kubernetes provider doesn't support loading config file.

* Added code level documentation comments.

* Moved all provider blocks to root environments modules.
Added comments.

* Removed commented code.
Rewrote some comments.