Switch to ko-based build for go applications
/kind feature
/area ci

This PR tackles down problem of managing multiple Dockerfiles where we usually only need to build small image with a single binary. This leverages the usage of ko as main build tool. To build an image/application it's needed to just add its import path to the .koapps.yaml

Then it will be automatically picked up by ko and built in next ProwJob run.