bug: workaround - static nodeport for gcsweb healthchecks (#9241)
/kind bug
/area ci

Currently GCP external LB doesn't support proper configuration of health checks for its load balancers, when health checks are served on different port.
This port must be exposed, so external LB can access it.
This change is a workaround, as it exposes healthcheck port and sets static nodePort on a cluster. Then custom BackendConfig uses this exposed port to manage healthchecks for a service.