Remove template.
The workflow for prow job manifests has changed. We no longer edit templates that get scaffolded, we now edit the job directly. To allow this, we need to remove the template first, so the prow job that monitors THIS PR does not complain about the difference between the prow definition file and the template.